distrobox-envs/ubuntu-python/tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "genesis-world>=0.2.1",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "ruff>=0.9.7",
    "torch>=2.6.0",
]
